+++ models/sellam2020/scripts/download-models.sh
@@ -33,3 +33,27 @@ if [ "$LARGE_512" = "true" ]; then
   unzip bleurt-large-512.zip
   rm bleurt-large-512.zip
 fi
+
+if [ "$BLEURT_20" = "true" ]; then
+  wget https://storage.googleapis.com/bleurt-oss-21/BLEURT-20.zip
+  unzip BLEURT-20.zip
+  rm BLEURT-20.zip
+fi
+
+if [ "$BLEURT_20_D12" = "true" ]; then
+  wget https://storage.googleapis.com/bleurt-oss-21/BLEURT-20-D12.zip
+  unzip BLEURT-20-D12.zip
+  rm BLEURT-20-D12.zip
+fi
+
+if [ "$BLEURT_20_D6" = "true" ]; then
+  wget https://storage.googleapis.com/bleurt-oss-21/BLEURT-20-D6.zip
+  unzip BLEURT-20-D6.zip
+  rm BLEURT-20-D6.zip
+fi
+
+if [ "$BLEURT_20_D3" = "true" ]; then
+  wget https://storage.googleapis.com/bleurt-oss-21/BLEURT-20-D3.zip
+  unzip BLEURT-20-D3.zip
+  rm BLEURT-20-D3.zip
+fi